server/requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
Django==1.8
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
django-pure-pagination==0.2.1
Pygments==1.6
logutils==0.3.3
django-extensions
webtest
django-webtest


django-picklefield
django-storages
boto
django-avatar
django-autofixture
django-filter

sendgrid-django 
django-activelink
sorl-thumbnail
django_markdown
markdown

# auth
django-allauth

# stripe
https://github.com/alialiev/dj-stripe/archive/master.zip

https://github.com/fernandoPardo/django-enhanced-cbv/archive/master.zip
https://github.com/yrik/django-allauth-multiple/archive/master.zip

#django-select2 User after fix of https://github.com/applegrew/django-select2/issues/121
# https://github.com/applegrew/django-select2/issues/166
#https://github.com/maximzxc/django-select2/archive/master.zip
#https://github.com/applegrew/django-select2/archive/master.zip 
https://github.com/applegrew/django-select2/archive/de37752ac63da4e180c63ad68c314f58d4f73bf6.zip

# See https://github.com/jezdez/django-constance/pull/96
#django-constance
https://github.com/jonzlin95/django-constance/archive/master.zip

#model_mommy
https://github.com/vandersonmota/model_mommy/archive/master.zip

# https://github.com/maraujop/django-crispy-forms/issues/263
#django-crispy-forms
https://github.com/maraujop/django-crispy-forms/archive/dev.zip

djangorestframework
django-cors-headers

django-debug-toolbar==1.3.2

dj-database-url==0.3.0
django-postgrespool==0.3.0
gunicorn==19.1.1
psycopg2==2.5.4
whitenoise==1.0.3
wsgiref==0.1.2

